规整和排除一些依赖

+++ lib_common/src/main/java/com/example/common/utils/file/FileUtil.kt
@@ -5,24 +5,29 @@ import android.content.Intent
 import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.PixelFormat
-import android.media.MediaScannerConnection
 import android.net.Uri
 import android.os.Build
 import android.provider.MediaStore
 import android.provider.Settings
+import android.text.TextUtils
 import androidx.core.content.FileProvider
 import com.example.base.utils.DateUtil
+import com.example.base.utils.LogUtil
 import com.example.common.constant.Constants
 import java.io.*
 import java.lang.ref.SoftReference
 import java.text.DecimalFormat
 import java.util.*
+import java.util.zip.ZipEntry
+import java.util.zip.ZipOutputStream
 
 /**
  * Created by WangYanBin on 2020/7/1.
  * 文件管理工具类
  */
 object FileUtil {
+
+    private const val TAG = "FileUtil"
 
     /**
      * 是否安装了XXX应用
@@ -43,7 +48,7 @@ object FileUtil {
      */
     @JvmStatic
     fun isRoot(): Boolean {
-        var file :File
+        var file: File
         val paths = arrayOf("/system/bin/", "/system/xbin/", "/system/sbin/", "/sbin/", "/vendor/bin/")
         try {
             for (element in paths) {
@@ -94,7 +99,8 @@ object FileUtil {
      * 删除文件
      */
     @JvmStatic
-    fun deleteFile(filePath: String) {
+    fun deleteFile(filePath: String?) {
+        if (TextUtils.isEmpty(filePath)) return
         val file = File(filePath)
         if (file.isFile && file.exists()) file.delete()
     }
@@ -116,6 +122,59 @@ object FileUtil {
             else if (file.isDirectory) deleteDirWithFile(file) //递规的方式删除文件夹
         }
         dir.delete() //删除目录本身
+    }
+
+    /**
+     * 将指定路径下的所有文件打成压缩包
+     * File fileDir = new File(rootDir + "/DCIM/Screenshots");
+     * File zipFile = new File(rootDir + "/" + taskId + ".zip");
+     *
+     * @param srcFilePath 要压缩的文件或文件夹路径
+     * @param zipFilePath 压缩完成的Zip路径
+     */
+    @JvmStatic
+    @Throws(Exception::class)
+    fun zipFolder(srcFilePath: String, zipFilePath: String) {
+        //创建ZIP
+        val outZip = ZipOutputStream(FileOutputStream(zipFilePath))
+        //创建文件
+        val file = File(srcFilePath)
+        //压缩
+        zipFiles(file.parent + File.separator, file.name, outZip)
+        //完成和关闭
+        outZip.finish()
+        outZip.close()
+    }
+
+    @Throws(Exception::class)
+    private fun zipFiles(folderPath: String, fileName: String, zipOutputSteam: ZipOutputStream?) {
+        LogUtil.e(TAG, " \n压缩路径:$folderPath\n压缩文件名:$fileName")
+        if (zipOutputSteam == null) return
+        val file = File(folderPath + fileName)
+        if (file.isFile) {
+            val zipEntry = ZipEntry(fileName)
+            val inputStream = FileInputStream(file)
+            zipOutputSteam.putNextEntry(zipEntry)
+            var len: Int
+            val buffer = ByteArray(4096)
+            while (inputStream.read(buffer).also { len = it } != -1) {
+                zipOutputSteam.write(buffer, 0, len)
+            }
+            zipOutputSteam.closeEntry()
+        } else {
+            //文件夹
+            val fileList = file.list()
+            //没有子文件和压缩
+            if (fileList.isEmpty()) {
+                val zipEntry = ZipEntry(fileName + File.separator)
+                zipOutputSteam.putNextEntry(zipEntry)
+                zipOutputSteam.closeEntry()
+            }
+            //子文件和递归
+            for (i in fileList.indices) {
+                zipFiles("$folderPath$fileName/", fileList[i], zipOutputSteam)
+            }
+        }
     }
 
     /**
